+++ polaris.shopify.com/constants.js
@@ -1,0 +1,6 @@
+// Not a TS file because our playroom.config.js needs to access it also, and can't understand ts imports.
+module.exports = {
+  playroom: {
+    baseUrl: '/playroom/',
+  },
+};