+++ polaris.shopify.com/playroom/components.ts
@@ -1,0 +1,2 @@
+export * from '@shopify/polaris';
+export * from '@shopify/polaris-icons';