+++ polaris.shopify.com/src/components/PlayroomLink/index.ts
@@ -1,0 +1,3 @@
+import PlayroomLink from './PlayroomLink';
+
+export default PlayroomLink;