+++ polaris.shopify.com/src/components/SandboxContainer/index.ts
@@ -1,0 +1,3 @@
+import SandboxContainer from './SandboxContainer';
+
+export default SandboxContainer;